refactor(cli): split mcp wrap helpers
